feat: add graphing of sensor data
