+++ package.json
@@ -57,6 +57,7 @@
     "redux": "^4.1.1",
     "redux-devtools-extension": "^2.13.9",
     "styled-components": "^5.3.1",
+    "swr": "^1.3.0",
     "typescript": "^4.4.3",
     "web-vitals": "^2.1.4"
   },
